requirements.txt
pytesseract==0.3.10
Pillow
Wand